Finished shopping fixed positions edit

+++ src/main/java/com/boparty/bopartycatering/Repos/IShoppingListItemRepos.java
@@ -3,5 +3,6 @@ package com.boparty.bopartycatering.Repos;
 import com.boparty.bopartycatering.Models.Order.ShoppingListItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IShoppingListItemRepos extends JpaRepository<ShoppingListItem, Integer> {
+public interface IShoppingListItemRepos extends JpaRepository<ShoppingListItem, Long> {
+//    ShoppingListItem findById(long id);
 }
